claude: refactor(forward/cli): move kaspa encoding to kaspaaddr package
Extract Kaspa address encoding into its own package for better
encapsulation. Renamed H256ToKaspaAddress to FromH256 since the
package name provides context.
