+++ x/forward/cli/kaspaaddr/kaspaaddr.go
@@ -1,4 +1,5 @@
-package cli
+// Package kaspaaddr provides Kaspa address encoding from H256 public keys.
+package kaspaaddr
 
 import (
 	"fmt"
@@ -7,16 +8,15 @@ import (
 	"github.com/btcsuite/btcd/btcutil/bech32"
 )
 
-// kaspa bech32m charset
-const kaspaCharset = "qpzry9x8gf2tvdw0s3jn54khce6mua7l"
+// charset is the Kaspa bech32m character set
+const charset = "qpzry9x8gf2tvdw0s3jn54khce6mua7l"
 
-// H256ToKaspaAddress converts a 32-byte H256 (Hyperlane recipient) to a Kaspa address.
+// FromH256 converts a 32-byte H256 (Hyperlane recipient) to a Kaspa address.
 // Kaspa uses a bech32-like encoding with key differences from standard bech32m:
 // - Separator is ':' instead of '1'
 // - 40-bit checksum (8 chars) instead of standard 30-bit (6 chars)
 // - Different generator polynomial and HRP expansion
-// These differences require custom checksum functions (kaspaChecksum, kaspaPolymod).
-func H256ToKaspaAddress(h256 []byte, mainnet bool) (string, error) {
+func FromH256(h256 []byte, mainnet bool) (string, error) {
 	if len(h256) != 32 {
 		return "", fmt.Errorf("invalid H256 length: expected 32, got %d", len(h256))
 	}
@@ -44,18 +44,17 @@ func H256ToKaspaAddress(h256 []byte, mainnet bool) (string, error) {
 	sb.WriteString(prefix)
 	sb.WriteString(":")
 	for _, v := range values {
-		sb.WriteByte(kaspaCharset[v])
+		sb.WriteByte(charset[v])
 	}
 	for _, v := range checksum {
-		sb.WriteByte(kaspaCharset[v])
+		sb.WriteByte(charset[v])
 	}
 
 	return sb.String(), nil
 }
 
-// kaspaChecksum computes the Kaspa bech32m 8-character checksum
-// Kaspa uses a 40-bit checksum (8 characters) instead of the standard 30-bit (6 characters)
-// It also uses a simplified HRP expansion (just hrp & 0x1f instead of standard bech32 expansion)
+// kaspaChecksum computes the Kaspa 8-character (40-bit) checksum.
+// This differs from standard bech32's 6-character (30-bit) checksum.
 func kaspaChecksum(hrp string, values []byte) []byte {
 	// Kaspa's simplified HRP expansion: just mask each byte to 5 bits
 	hrpExp := make([]byte, 0, len(hrp)+1)
@@ -79,10 +78,8 @@ func kaspaChecksum(hrp string, values []byte) []byte {
 	return checksum
 }
 
-// kaspaPolymod computes the Kaspa 40-bit polymod
-// This is different from standard bech32 which uses 30-bit
+// kaspaPolymod computes the Kaspa 40-bit polymod (vs standard bech32's 30-bit).
 func kaspaPolymod(values []byte) uint64 {
-	// Generator polynomial coefficients for 40-bit checksum
 	gen := []uint64{
 		0x98f2bc8e61,
 		0x79b76d99e2,
